OPHJOD-1967: Use design system's breadcrumb component
