+++ src/components/MainLayout/MainLayout.tsx
@@ -1,6 +1,7 @@
-import { tidyClasses } from '@jod/design-system';
+import { useBreadcrumbItems } from '@/hooks/useBreadcrumbItems';
+import { Breadcrumb, tidyClasses } from '@jod/design-system';
 import React from 'react';
-import { Breadcrumb } from '../Breadcrumb/Breadcrumb';
+import { NavLink } from 'react-router';
 
 interface MainLayoutProps {
   children: React.ReactNode;
@@ -9,12 +10,13 @@ interface MainLayoutProps {
 }
 
 export const MainLayout = ({ children, navChildren, asideChildren }: MainLayoutProps) => {
+  const breadcrumbItems = useBreadcrumbItems();
   return (
     <div
       className="mx-auto grid w-full max-w-[1140px] grow grid-cols-3 gap-6 px-5 pb-6 pt-5 sm:px-6 print:p-0 auto-rows-max"
       data-testid="main-layout"
     >
-      <Breadcrumb />
+      <Breadcrumb items={breadcrumbItems} serviceVariant="ohjaaja" LinkComponent={NavLink} />
 
       {(navChildren || asideChildren) && (
         <aside
